requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
importlib-metadata==4.5.0
Markdown==3.3.4
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
zipp==3.4.1
